\name{addSets}
\alias{addSets}
\title{addSets}
\description{Prepare new set of parameter vectors for \code{kproject}}
\usage{addSets(kproject, nStart = 1, nSets = 500, seed = 100)}
\arguments{
  \item{kproject}{project to prepare sets for}
  \item{nStart}{start index of the set}
  \item{nSets}{number of the set}
  \item{seed}{random generator seed to expand existing set in the project use \code{kproject$seed}}
}



\author{Anatoly Sorokin}





