\name{makeBruttoStr}
\alias{makeBruttoStr}
\title{makeBruttoStr}
\description{create string representation of brutto formula from kappa complex definition string  }
\usage{makeBruttoStr(kappa)}
\arguments{
  \item{kappa}{kappa complex definition string}
}



\author{Anatoly Sorokin}





