\name{makeIGraph}
\alias{makeIGraph}
\title{makeIGraph}
\description{function converts kappa string, defining the complex structure into iGraph graph}
\usage{makeIGraph(kappa)}
\arguments{
  \item{kappa}{complex definition in kappa language}
}

\value{simplified iGraph graph of the complex}

\author{Anatoly Sorokin}





