\name{read.observables.kproject}
\alias{read.observables.kproject}
\title{read observables kproject}
\description{function to read and parse KaSim simulation results}
\usage{\method{read.observables}{kproject}(kproject, dir)}
\arguments{
  \item{kproject}{project to handle}
  \item{dir}{name of the folder with project simulation results}
}

\value{object of \code{kprojectSimExp} class that stores both project and simulation experiment results}

\author{Anatoly Sorokin}





