\name{read.observables}
\alias{read.observables}
\title{read observables}
\description{function to read and parse KaSim simulation results}
\usage{read.observables(file)}
\arguments{
  \item{file}{name of the observable file}
}

\value{data frame of observables}

\author{Anatoly Sorokin}





