\name{read.sim.folder}
\alias{read.sim.folder}
\title{read sim folder}
\description{utulity function to read content of one \code{pset} folder}
\usage{read.sim.folder(file, aggregate = FALSE, block.size = 10)}
\arguments{
  \item{file}{location of the folder to read}
  \item{aggregate}{wether aggregate data on the fly or return raw dataset}
  \item{block.size}{number of simulations to aggregate}
}

\value{data from the folder}

\author{Anatoly Sorokin}





