\name{replaceValue}
\alias{replaceValue}
\title{replaceValue}
\description{Replace named parameter with value}
\usage{replaceValue(pname, value, lines)}
\arguments{
  \item{pname}{name of the property to substitute with value}
  \item{value}{value to replace with}
  \item{lines}{string vector to make substitution in}
}



\author{Anatoly Sorokin}





