\name{rkappa-package}
\alias{rkappa-package}
\docType{package}
\title{Kappa simulation Project}
\description{The package provides a infrastructure required to simulate kappa model with various parameter sets and analyse results of such simulations.}
\details{
\tabular{ll}{Package: \tab rkappa\cr
Type: \tab Package\cr
Title: \tab Kappa simulation Project\cr
Version: \tab 1.0.130121\cr
Date: \tab 2013-01-21\cr
Author: \tab Anatoly Sorokin\cr
Maintainer: \tab Anatoly Sorokin <lptolik@gmail.com>\cr
License: \tab GPL (>= 2)\cr
Copyright: \tab Anatoly Sorokin 2012-2013\cr
Depends: \tab R (>= 2.14.0),randtoolbox,gdata,igraph,plyr\cr
Suggests: \tab lattice,sensitivity\cr}
}
\author{Anatoly Sorokin}

\keyword{ package }


