\name{rkappa-package}
\alias{rkappa-package}
\alias{rkappa}
\docType{package}
\title{Kappa simulation Project}
\description{The package provides a infrastructure required to simulate kappa model with various parameter sets and analyse results of such simulations.}
\details{
\tabular{ll}{Package: \tab rkappa\cr
Type: \tab Package\cr
Title: \tab Kappa simulation Project\cr
Version: \tab 1.1.150303\cr
Date: \tab 2015-03-05\cr
Author: \tab Anatoly Sorokin\cr
Maintainer: \tab Anatoly Sorokin <lptolik@gmail.com>\cr
License: \tab GPL (>= 2)\cr
Copyright: \tab Anatoly Sorokin 2012-2015\cr
VignetteBuilder: \tab knitr\cr
Depends: \tab R (>= 3.0.0),randtoolbox,gdata,igraph,plyr,futile.logger\cr
Suggests: \tab lattice,sensitivity,RColorBrewer,knitr (>= 1.6),xtable\cr}
}
\author{Anatoly Sorokin}

\keyword{ package }


